feat(EatUp_Client): add Picasso library for image loading and error and placeholder images
This commit adds the Picasso library to the project to load images from a URL. It also adds two new images, ic_error_load.png and ic_load_image.png, to be used as error and placeholder images, respectively.

feat(EatUp_Client): use Picasso to load images in ProductAdapter and ResumeAdapter

This commit replaces the hardcoded image resource in the ProductAdapter and ResumeAdapter with a call to Picasso to load the image from a URL. It also adds padding to the image in the item_product.xml layout.

style(EatUp_Client): adjust layout in item_product.xml

This commit adjusts the layout in the item_product.xml file to add padding and adjust the width of the add product button and the quantity view.